update build and steps
